Simplify SQL generation for RQL's _eq operator
Currently in RQL, `{"column": {"_eq": null}}` automatically checks if
the column `IS NULL` (unlike GraphQL). The SQL expression we generate
for `{"column": {"_eq": <exp>}}` is

```
"column" = <exp> OR ("column" IS NULL AND <exp> IS NULL)
```

This is so that the generated SQL statement is the same whether `<exp>`
is `null` or `<exp>` is some value `v`. This helps when we are using
prepared statements. However, given that RQL is also used for specifying
permissions, seeing such SQL is confusing for our users. `{"column":
{"_eq": null}}` now translates to `column IS NULL` and `{"column":
{"_eq": 1}}` translates to `column = 1`